[benchmarks] rewrite HTML visualization to be js-native
This patch changes how the website to visualize benchmark data
is generated. Before, we used matplotlib and mpld3 to prepare
a huge HTML file where we attempted to generate javascript from
matplotlib-like charts. This was very painful, and half of the
features of matplotlib didn't work properly, resulting in
many workarounds in the code.

With this patch, all of the logic, except for simple data processing,
is now done inside of a dedicated javascript file, and the visualization
itself is handled through Chart.js. This results in both smaller files
that are quicker to load and more visually pleasing website.

What this also enabled is decoupling of the HTML file from the
charts data that previously used to be embedded. The new HTML
website now has the option of adding and removing benchmark runs.
This is in preparation of allowing per-PR visualization.